// src/app/claim-processing/manual-claim-processing/process-claim-medsupp-chargeb/model/screen-save-area.model.ts
import {ScreenChargeLine} from './screen-charge-line.model';

/**
 * Model class ScreenSaveArea
 * Path: screenbean/procclmmedsupptbchrgb
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmmedsupptbchrgb::ScreenSaveArea
 * Legacy Mapping: SCREEN-SAVE-AREA
 */
export class ScreenSaveArea {
  filler1 = '';
  mapCommandLine = '';
  filler2 = '';
  m9Icd = '';
  m9Icd2 = '';
  m9Icd3 = '';
  screenChargeLines: ScreenChargeLine[] = [];
  screenAfterTermPlan = '';
  filler14 = '';
  screenExclPlan1 = '';
  filler15 = '';
  screenExclPlan2 = '';
  filler16 = '';
  screenErr1 = '';
  m9hcmnd = '';
  m9hmem1 = '';
  m9hmem2 = '';
  m9hmem3 = '';
  m9hcno1 = '';
  m9hcno2 = '';
  m9hcno3 = '';
  m9hcno4 = '';
  m9hcno5 = '';
  m9hcno6 = '';
  m9hcom1 = '';
  m9hcom2 = '';
  m9hname = '';
  m9hpln1 = '';
  m9heff1 = '';
  m9htrm1 = '';
  m9hmemn = '';
  m9hassc = '';
  m9hpln2 = '';
  m9heff2 = '';
  m9htrm2 = '';
  m9hdob = '';
  m9hpln3 = '';
  m9heff3 = '';
  m9htrm3 = '';
}